src/server/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
psycopg2-binary==2.9.6
redis==5.0.1
black==23.12.0
Flask-Bcrypt==1.0.1
boto3==1.34.2
